+++ src/humbledownloader.py
@@ -1,5 +1,4 @@
-from model.game import HumbleGame, TroveGame, Subproduct
-from model.download import DownloadStructItem, SubproductDownload, TroveDownload
+from model.download import DownloadStructItem, SubproductDownload
 from consts import CURRENT_BITNESS, BITNESS